Refactor clients Appwrite
Moved appwrite clients in separate files
