+++ app/(auth)/sign-in.tsx
@@ -5,8 +5,8 @@ import { View, Text, ScrollView, Dimensions, Alert, Image } from "react-native";
 
 import { images } from "../../constants";
 import { CustomButton, FormField } from "../../components";
-import { getCurrentUser, signIn } from "../../lib/appwrite";
 import { useGlobalContext } from "../../context/GlobalProvider";
+import { getCurrentUser, signIn } from "../../lib/clients/user";
 
 const SignIn = () => {
   const { setUser, setIsLogged } = useGlobalContext();
